# Conflicts:
#	src/service/Filter/test/UnitTests/VSS.Productivity3D.Filter.Tests/BoundaryExecutorTests.cs
#	src/service/Filter/test/UnitTests/VSS.Productivity3D.Filter.Tests/FilterExecutorTests.cs
#	src/service/Filter/test/UnitTests/VSS.Productivity3D.Filter.Tests/FilterFilenameUtilTests.cs

+++ src/service/TRex/src/netstandard/VSS.TRex.Analytics/CMVChangeStatistics/CMVChangeStatisticsAggregator.cs
@@ -70,7 +70,7 @@ namespace VSS.TRex.Analytics.CMVChangeStatistics
                 if (cmvValue.PreviousMeasuredCMV == CellPassConsts.NullCCV)
                   cmvChangeValue = MAX_PERCENTAGE_VALUE;
                 else
-                  cmvChangeValue = (cmvValue.MeasuredCMV - cmvValue.PreviousMeasuredCMV) / cmvValue.PreviousMeasuredCMV * MAX_PERCENTAGE_VALUE;
+                  cmvChangeValue = (double)(cmvValue.MeasuredCMV - cmvValue.PreviousMeasuredCMV) / (double)cmvValue.PreviousMeasuredCMV * MAX_PERCENTAGE_VALUE;
 
                 IncrementCountOfTransition(cmvChangeValue); // CMV Change Detail is counted here...
               }
